requirement.txt
streamlit
numpy
pandas
scikit-learn
matplotlib
plotly

librosa
pydub
sounddevice

openai-whisper

spacy
nltk
sentence-transformers

opencv-python
mediapipe

sqlite-utils
streamlit
numpy
pandas
scikit-learn
matplotlib
plotly

librosa
pydub
sounddevice

openai-whisper

spacy
nltk
sentence-transformers

opencv-python
mediapipe

sqlite-utils
